mnist-pytorch/requirements.txt
torch
torchvision==0.20.0
s3fs==2025.3.0
python-dotenv==1.0.1
mlflow>=2.19.0
pelicanfs==1.1.2